MMCQ.quantize returns exactly the number of colors specified
The inner function used by MMCQ.quantize was counting the number of colors
added to the palette, starting from 1. This only works when the PQueue passed
contains only 1 VBox; but the iter function is called twice, and the second
time, the VBox contains more than one color. This means that in some
circumstances it was counting the number of colors added incorrectly, and so
MMCQ.quantize (and ColorThief.getPalette) would sometimes return the wrong
number of colors. This commit changes iter to start counting at the actual
number of colors in the PQueue, so that it always returns the number of
colors requested.